extract item name resolving logic into own class
